Update function signature
