+++ update-playlist-storage-v2/env.dev.yaml
@@ -1,0 +1,3 @@
+---
+  BUCKET: chirpradiodev-public
+  FILE_NAME: playlist-test.json
