--- update-playlist-storage-v2/lib/index.js
@@ -1,9 +1,0 @@
-const { formatTrack } = require("./formatTrack");
-const { loadJson } = require("./loadJson");
-const { updateNowPlaying } = require("./updateNowPlaying");
-
-module.exports = {
-  formatTrack,
-  loadJson,
-  updateNowPlaying,
-};
