+++ update-playlist-storage-v2/package.json
@@ -6,8 +6,8 @@
   "author": "Steve Lund <lund5000@gmail.com>",
   "license": "MIT",
   "scripts": {
-    "start": "functions-framework --target=updateNowPlaying --port=8081",
-    "deploy:dev": "gcloud functions deploy updatePlaylistStorage --project chirpradiodev --runtime=nodejs16 --env-vars-file env.dev.yaml --trigger-topic playlist-track-processed",
+    "start": "yaml2json env.dev.yaml >.env.json && ./node_modules/.bin/env-cmd -f ./.env.json functions-framework --target=updatePlaylistStorageV2 --port=8082",
+    "deploy:dev": "gcloud functions deploy updatePlaylistStorageV2 --project chirpradiodev --runtime=nodejs16 --env-vars-file env.dev.yaml --trigger-topic playlist-track-processed",
     "deploy": "gcloud functions deploy updatePlaylistStorage --project chirpradio-hrd --runtime=nodejs10 --trigger-http --allow-unauthenticated",
     "test": "jest",
     "lint": "npx eslint ."
@@ -19,8 +19,10 @@
     "luxon": "^3.4.4"
   },
   "devDependencies": {
+    "env-cmd": "^10.1.0",
     "eslint": "^8.56.0",
     "jest": "^27.5.1",
-    "prettier-eslint": "^16.3.0"
+    "prettier-eslint": "^16.3.0",
+    "yaml2json": "^1.0.2"
   }
 }
